refactor(aarch64): split vmgen.rs into vmgen/ submodules mirroring x86
Pure code movement, no behavior change. Mirrors `arch/x86_64/vmgen{.rs,/}`:
`vmgen.rs` stays the module file (VM construction, the arithmetic/control/
array/hash/index/defined opcode handlers, BOP de-opt) and now declares the same
four submodules x86 does — `definition` (class/method def opcodes),
`pub mod init_method`, `method_call` (send / yield / cache save), and
`variables` (const / ivar / cvar / gvar / dvar access opcodes) — each handler
placed to match its x86 counterpart. The moved `a64_op_*` handlers were already
`pub(in crate::codegen)`, so visibility is unchanged.

Verified: builds warning-free; VM-tier opcode smoke tests pass. Full suite in
the accompanying run.
